extract defaultState into own module; fix flow for compact-toggle-tests
